test_requirements.txt
pytest==8.3.4
pytest-django==4.9.0
pytest-asyncio==0.25.0
pytest-mock==3.14.0
pytest-freezegun==0.4.2
factory-boy==3.3.1
Faker==33.0.0